separated popup into a module
